# Add zf_log to performance tests