Adding [zf_log](https://github.com/wonder-mice/zf_log) to performance tests. On my machine (MacBook Pro 2.8GHz Core i7) I'm getting following results (real time, best of 3):
```
        1 thread   10 threads   100 threads
zf_log   0.466s     2.711s       2.608s
spdlog   0.747s     5.121s       5.376s
```